streamed_microphone_transcription/requirements.txt
pyaudio
openai-whisper
speechbrain
torchvision
SpeechRecognition